refactor(model-providers): delete legacy seed file, finalize module pattern
Drop the transitional shim now that every built-in lives in a module:

  - Delete apps/api/src/services/oauth-model-providers/registry.ts (was
    re-exporting from model-providers/registry.ts + ../codex via a
    back-compat alias since PR 2; LEGACY_PROVIDERS has been empty
    since PR 6).
  - boot.ts: drop seedLegacyModelProviders() call — module contributions
    are now the sole source of truth, populated via getModuleModelProviders().
  - oauth-flow.ts, token-resolver.ts: import decodeCodexJwtPayload from
    ../../modules/codex/index.ts (its real owner) instead of the legacy
    re-export.
  - refresh-worker.ts: import listModelProviders from the canonical
    runtime registry.
  - run-launcher/pi.ts: same — codex helper at its real path.
  - test/helpers/app.ts + tests: drop seed call + MODEL_PROVIDERS assertion.
  - CLAUDE.md: MODULES default is now oidc,webhooks,core-providers
    (codex stays opt-in). MODEL_PROVIDERS_DISABLED doc loses the
    claude-code reference and the PR #397 footnote.
