+++ apps/api/src/services/run-launcher/pi.ts
@@ -39,7 +39,7 @@ import type { SinkCredentials } from "../../lib/mint-sink-credentials.ts";
 
 import { getEnv } from "@appstrate/env";
 import { isOAuthModelProvider } from "../model-providers/registry.ts";
-import { decodeCodexJwtPayload } from "../oauth-model-providers/registry.ts";
+import { decodeCodexJwtPayload } from "../../modules/codex/index.ts";
 import type { LlmProxyConfig, LlmProxyOauthConfig } from "@appstrate/core/sidecar-types";
 
 /** Terminal state reported back to the caller once the container has exited. */
